initial setup for modifications to Lattice and Math Layers
