+++ src/core/include/lattice/matrix-lattice-impl.h
