--- src/core/lib/lattice/hal/default/dcrtpoly-impl.cpp
@@ -1,59 +1,0 @@
-//==================================================================================
-// BSD 2-Clause License
-//
-// Copyright (c) 2014-2022, NJIT, Duality Technologies Inc. and other contributors
-//
-// All rights reserved.
-//
-// Author TPOC: contact@openfhe.org
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-// 1. Redistributions of source code must retain the above copyright notice, this
-//    list of conditions and the following disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright notice,
-//    this list of conditions and the following disclaimer in the documentation
-//    and/or other materials provided with the distribution.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//==================================================================================
-
-/*
-  Implementation of the integer lattice using double-CRT representations
- */
-
-#include "lattice/hal/default/dcrtpoly.cpp"    // NOLINT
-#include "lattice/elemparams.cpp"              // NOLINT
-#include "lattice/ildcrtparams.cpp"            // NOLINT
-#include "lattice/poly.cpp"                    // NOLINT
-#include "math/binaryuniformgenerator.cpp"     // NOLINT
-#include "math/discretegaussiangenerator.cpp"  // NOLINT
-#include "math/discreteuniformgenerator.cpp"   // NOLINT
-#include "math/ternaryuniformgenerator.cpp"    // NOLINT
-
-// This creates all the necessary class implementations for DCRTPoly
-
-namespace lbcrypto {
-
-template class ILDCRTParams<M2Integer>;
-template class DCRTPolyImpl<M2Vector>;
-template class ILDCRTParams<M4Integer>;
-template class DCRTPolyImpl<M4Vector>;
-
-#ifdef WITH_NTL
-template class ILDCRTParams<M6Integer>;
-template class DCRTPolyImpl<M6Vector>;
-#endif
-
-}  // namespace lbcrypto
