+++ src/core/lib/lattice/lattice.cpp
