+++ src/core/lib/lattice/trapdoor-dcrtpoly.cpp
