+++ src/core/lib/lattice/trapdoor-poly.cpp
